+++ string.js
@@ -1,4 +1,4 @@
-import dedent from './node_modules/string-dedent/dist/dedent.mjs';
+import dedent from './assets/dedent.js';
 
 if (! (String.dedent instanceof Function)) {
 	String.dedent = dedent;